feat: refactored everything to simplify; use cuda 12.1, runpod 1.7.9
