+++ Dockerfile
@@ -1,20 +1,37 @@
-# Base image
-FROM runpod/base:0.4.2-cuda11.8.0
+# base image with cuda 12.1
+FROM nvidia/cuda:12.1.1-cudnn8-runtime-ubuntu22.04
 
-ENV HF_HUB_ENABLE_HF_TRANSFER=0
+# install python 3.11 and pip
+ENV DEBIAN_FRONTEND=noninteractive
+RUN apt-get update && apt-get install -y --no-install-recommends \
+    software-properties-common \
+    && add-apt-repository ppa:deadsnakes/ppa \
+    && apt-get update && apt-get install -y --no-install-recommends \
+    python3.11 \
+    python3.11-venv \
+    python3-pip \
+    git \
+    && rm -rf /var/lib/apt/lists/*
 
-# Install Python dependencies (Worker Template)
-COPY builder/requirements.txt /requirements.txt
-RUN python3.11 -m pip install --upgrade pip && \
-    python3.11 -m pip install --upgrade -r /requirements.txt --no-cache-dir && \
-    rm /requirements.txt
+# set python3.11 as the default python
+RUN ln -sf /usr/bin/python3.11 /usr/local/bin/python && \
+    ln -sf /usr/bin/python3.11 /usr/local/bin/python3
 
-# Cache Models
-COPY builder/cache_models.py /cache_models.py
-RUN python3.11 /cache_models.py && \
-    rm /cache_models.py
+# install uv
+RUN pip install uv
 
-# Add src files (Worker Template)
-ADD src .
+# create venv
+ENV PATH="/.venv/bin:${PATH}"
+RUN uv venv --python 3.11 /.venv
 
-CMD python3.11 -u /rp_handler.py
+# install dependencies
+RUN uv pip install torch --extra-index-url https://download.pytorch.org/whl/cu121 diffusers transformers accelerate safetensors xformers==0.0.23 runpod numpy==1.26.3 scipy triton huggingface-hub hf_transfer hf_xet setuptools
+
+# copy files
+COPY download_weights.py schemas.py handler.py test_input.json /
+
+# download the weights from hugging face
+RUN python /download_weights.py
+
+# run the handler
+CMD python -u /handler.py
