+++ download_weights.py
@@ -1,13 +1,15 @@
-# builder/model_fetcher.py
-
 import torch
-from diffusers import StableDiffusionXLPipeline, StableDiffusionXLImg2ImgPipeline, AutoencoderKL
+from diffusers import (
+    StableDiffusionXLPipeline,
+    StableDiffusionXLImg2ImgPipeline,
+    AutoencoderKL,
+)
 
 
 def fetch_pretrained_model(model_class, model_name, **kwargs):
-    '''
+    """
     Fetches a pretrained model from the HuggingFace model hub.
-    '''
+    """
     max_retries = 3
     for attempt in range(max_retries):
         try:
@@ -15,29 +17,35 @@ def fetch_pretrained_model(model_class, model_name, **kwargs):
         except OSError as err:
             if attempt < max_retries - 1:
                 print(
-                    f"Error encountered: {err}. Retrying attempt {attempt + 1} of {max_retries}...")
+                    f"Error encountered: {err}. Retrying attempt {attempt + 1} of {max_retries}..."
+                )
             else:
                 raise
 
 
 def get_diffusion_pipelines():
-    '''
+    """
     Fetches the Stable Diffusion XL pipelines from the HuggingFace model hub.
-    '''
+    """
     common_args = {
         "torch_dtype": torch.float16,
         "variant": "fp16",
-        "use_safetensors": True
+        "use_safetensors": True,
     }
 
-    pipe = fetch_pretrained_model(StableDiffusionXLPipeline,
-                                  "stabilityai/stable-diffusion-xl-base-1.0", **common_args)
+    pipe = fetch_pretrained_model(
+        StableDiffusionXLPipeline,
+        "stabilityai/stable-diffusion-xl-base-1.0",
+        **common_args,
+    )
     vae = fetch_pretrained_model(
         AutoencoderKL, "madebyollin/sdxl-vae-fp16-fix", **{"torch_dtype": torch.float16}
     )
-    print("Loaded VAE")
-    refiner = fetch_pretrained_model(StableDiffusionXLImg2ImgPipeline,
-                                     "stabilityai/stable-diffusion-xl-refiner-1.0", **common_args)
+    refiner = fetch_pretrained_model(
+        StableDiffusionXLImg2ImgPipeline,
+        "stabilityai/stable-diffusion-xl-refiner-1.0",
+        **common_args,
+    )
 
     return pipe, refiner, vae
